refactor(tool): extract snapshot export into a shared module
`Filecoin.ChainExportDiff` (in the RPC layer) called `do_export` directly
from the `archive_cmd` CLI subcommand module, coupling the RPC layer to a
CLI subcommand. Move `do_export` and its `build_output_path` helper into a
dedicated `crate::tool::export` module that both the CLI subcommands and
the RPC method import. No behaviour change.

Closes #6096
