+++ src/tool/mod.rs
@@ -1,5 +1,6 @@
 // Copyright 2019-2026 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
+pub mod export;
 pub mod main;
 pub mod offline_server;
 pub mod subcommands;
